requirements.txt
discord.py
python-dotenv
translate
langdetect
certifi